Add shm_object, flags, and tracing code.
The concept:

- shm_object is the interface between the
platform's shared memory API and the
shm_ptr<> we'll want to use.

- flags basically maps to Linux fcntl, but
doesn't tie pshm to the POSIX o_flags.

- allocation_tracer and related macros
provide tools to help see what's going on,
and specifically to be able to dump details
at the end of unit tests.

- PSHM_TRACING option controls whether
or not that code is compiled in.