Add the top level shm_ptr.
Semantics are similar to unique_ptr, as
that lends itself pretty well to control over
the shm_object and how its life cycle
is managed. Ditto for the virtual memory
systems on most platforms with memory
mapped file I/O.

For my use cases this is ideal, because
wanting 'another' shm_ptr can be solved
by creating a new one with the same
name. But simply copying ala shared_ptr,
doesn't lend itself so well given the flags
parameter of shm_object.

Particularly in the case of flags like O_EXCL,
where you're gonna make a blank stare
eventually. Whether you expect that copy
to fail because O_EXCL on the open or to
succeed because copying shm_ptr.