Handle shm_object move ctor/assignment.
- Deletes copy ctor/assignment.

- Implements move ctor/assignment.

My internal debate: explicitly delete
the move operations, so that it is harder
for the in-experienced to shoot themselves
with differences between move a pointer
and move an actual value, etc.

Since there's really know clean way to make
sure a user doesn't do something stupid
and the prevent us and users from doing
stupid in extension classes, I'm making
this an exercise in having working move
support for the sake of testing.

Debate: whether moving should clear
the name/flags/size/offset fields.

Leaving them as is, as shm_object doesn't
have a nothing there concept the way that
shm_ptr<> does.

In the end, shm_object is not really intended
to be used directly. It's intended more for
shm_ptr and other facilities like it.

But overall I feel this needs deeper thought.