Consolidate shm_object selection.
Since the library is only intended to be
using one native type, I don't see a point
in fancier selection. No reason to fall back
from POSIX shm to raw fd/mmap or SYSV.

Ditto, AFAICT the comparable facilities on
Windows NT are file mappings and file views.
Not a broad range of API of the decade
problem.